feature: area of slices and volume derivation
Uses
[skimage.measure.regionprops](https://scikit-image.org/docs/stable/api/skimage.measure.html#skimage.measure.regionprops)
on...

- labelled objects in individual slices of 3D array
- labelled objects in 3D array as a whole (i.e. area == volume)
- uses `pixel_to_nm_scaling` for `spacing` so units are correct.

Currently the `max_feret_diameter` is causing issues with the `sample1` and `sample2` test fixtures when applied to 3D
arrays, I've had a brief look at this and thought it may be down to watershedding not providing a background as `0` but
am unsure. I found a related [thread](https://github.com/scikit-image/scikit-image/issues/7822) where large amounts of
memory were trying to be allocated during a call to `measure.regionprops()` but not definitive answer.